+++ Assets/0.Scripts/ScriptableData/MainData.cs
@@ -1,11 +1,12 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System.Numerics;
 
 [CreateAssetMenu(fileName = "MainData", menuName = "ScriptableObjects/MainData")]
 public class MainData : ScriptableObject
 {
-    public int gold;
+    public BigInteger gold;
     public int stageLevel;
     // 인벤토리 정보
